requirements.txt
streamlit
pdfplumber
sentence-transformers
openai
python-dotenv
